requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
qrcode[pil]==7.4.2
SQLAlchemy==2.0.23
werkzeug==3.0.1
python-dotenv==1.0.0

